refactor: consolidate useAuthState hook and AuthState type
• Merged duplicate useAuthState implementations into single unified hook
• Consolidated AuthState type definition in hooks/useAuthState.tsx
• Combined features from both versions (analytics, callbacks, state checks)
• Updated all imports to use unified hook via hooks/index.ts export
• Removed duplicate file components/auth/hooks/useAuthState.tsx
• Maintains backward compatibility with enhanced functionality
