+++ hooks/index.ts
@@ -1,0 +1,5 @@
+// Central export for all application hooks
+
+// Authentication hooks
+export { useAuthState } from "./useAuthState";
+export type { AuthState, AuthFlow, UseAuthStateOptions, UseAuthStateReturn } from "./useAuthState";
