+++ hooks/useAuthState.tsx
@@ -3,16 +3,24 @@
 import { useState, useCallback } from "react";
 import { usePostHog } from "posthog-js/react";
 
-export type AuthFlow = "signup" | "login" | "reset" | "verify" | "forgot";
+// Single source of truth for AuthState type
 export type AuthState = "loading" | "form" | "success" | "error";
+export type AuthFlow = "signup" | "login" | "reset" | "verify" | "forgot";
 
-interface UseAuthStateOptions {
-  flow: AuthFlow;
+export interface UseAuthStateOptions {
+  /** Initial state to start with */
+  initialState?: AuthState;
+  /** Auth flow type for analytics */
+  flow?: AuthFlow;
+  /** Callback when state changes */
+  onStateChange?: (newState: AuthState) => void;
+  /** Callback on success */
   onSuccess?: () => void;
+  /** Callback on error */
   onError?: (error: string) => void;
 }
 
-interface UseAuthStateReturn {
+export interface UseAuthStateReturn {
   // State
   state: AuthState;
   isSubmitting: boolean;
@@ -29,72 +37,113 @@ interface UseAuthStateReturn {
   finishSubmitting: () => void;
   handleSuccess: () => void;
   handleError: (error: string | Error) => void;
+  reset: () => void;
+
+  // State checks
+  isLoading: boolean;
+  isSuccess: boolean;
+  isError: boolean;
 
   // Analytics
   trackEvent: (eventName: string, properties?: Record<string, any>) => void;
 }
 
+/**
+ * Unified hook for managing authentication flow states.
+ * Provides a consistent way to handle loading, form, success, and error states
+ * with optional analytics tracking and callbacks.
+ */
 export function useAuthState({
+  initialState = "form",
   flow,
+  onStateChange,
   onSuccess,
   onError,
-}: UseAuthStateOptions): UseAuthStateReturn {
-  const [state, setState] = useState<AuthState>("loading");
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [error, setError] = useState<string | null>(null);
+}: UseAuthStateOptions = {}): UseAuthStateReturn {
+  const [state, setStateInternal] = useState<AuthState>(initialState);
+  const [isSubmitting, setIsSubmittingInternal] = useState(false);
+  const [error, setErrorInternal] = useState<string | null>(null);
   const posthog = usePostHog();
 
   const trackEvent = useCallback(
     (eventName: string, properties?: Record<string, any>) => {
-      if (posthog) {
+      if (posthog && flow) {
         posthog.capture(eventName, {
           auth_flow: flow,
           ...properties,
         });
+      } else if (posthog) {
+        posthog.capture(eventName, properties);
       }
     },
     [posthog, flow]
   );
 
+  const setState = useCallback(
+    (newState: AuthState) => {
+      setStateInternal(newState);
+      if (onStateChange) {
+        onStateChange(newState);
+      }
+    },
+    [onStateChange]
+  );
+
   const setSubmitting = useCallback((submitting: boolean) => {
-    setIsSubmitting(submitting);
+    setIsSubmittingInternal(submitting);
     if (submitting) {
-      setError(null);
+      setErrorInternal(null);
     }
   }, []);
 
+  const setError = useCallback((error: string | null) => {
+    setErrorInternal(error);
+  }, []);
+
   const clearError = useCallback(() => {
-    setError(null);
+    setErrorInternal(null);
   }, []);
 
   const startSubmitting = useCallback(() => {
     setSubmitting(true);
-    trackEvent(`${flow}_attempt_start`);
+    if (flow) {
+      trackEvent(`${flow}_attempt_start`);
+    }
   }, [setSubmitting, trackEvent, flow]);
 
   const finishSubmitting = useCallback(() => {
     setSubmitting(false);
-  }, []);
+  }, [setSubmitting]);
 
   const handleSuccess = useCallback(() => {
     setState("success");
-    setIsSubmitting(false);
-    setError(null);
-    trackEvent(`${flow}_success`);
+    setIsSubmittingInternal(false);
+    setErrorInternal(null);
+    if (flow) {
+      trackEvent(`${flow}_success`);
+    }
     onSuccess?.();
-  }, [flow, trackEvent, onSuccess]);
+  }, [flow, trackEvent, onSuccess, setState]);
 
   const handleError = useCallback(
     (err: string | Error) => {
       const errorMessage = err instanceof Error ? err.message : err;
       setState("error");
-      setIsSubmitting(false);
-      setError(errorMessage);
-      trackEvent(`${flow}_error`, { error_message: errorMessage });
+      setIsSubmittingInternal(false);
+      setErrorInternal(errorMessage);
+      if (flow) {
+        trackEvent(`${flow}_error`, { error_message: errorMessage });
+      }
       onError?.(errorMessage);
     },
-    [flow, trackEvent, onError]
+    [flow, trackEvent, onError, setState]
   );
+
+  const reset = useCallback(() => {
+    setState(initialState);
+    setErrorInternal(null);
+    setIsSubmittingInternal(false);
+  }, [initialState, setState]);
 
   return {
     // State
@@ -113,6 +162,12 @@ export function useAuthState({
     finishSubmitting,
     handleSuccess,
     handleError,
+    reset,
+
+    // State checks
+    isLoading: state === "loading",
+    isSuccess: state === "success",
+    isError: state === "error",
 
     // Analytics
     trackEvent,
